refactor(core): home decode, punctuation, and VAD tests
Turn decode.rs, punctuation.rs, and vad.rs into directory modules so the
unit tests sit next to the impl instead of doubling the file. Public
paths are unchanged.
